use double values for font sizes. scale X11 to screen dpi, and png/jpeg to res/72 dpi. correct/update documentation.
git-svn-id: https://svn.r-project.org/R/trunk@44553 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/modules/X11/devX11.h
@@ -91,7 +91,7 @@ typedef struct {
     int canvas;				/* Canvas */
     int fontface;			/* Typeface */
     int fontsize;			/* Size in points */
-    int pointsize;			/* Size in points */
+    double pointsize;			/* Size in points */
     char basefontfamily[500];           /* Initial font family */
 
     /* X11 Driver Specific */
